fix(routes): cast jsonb_array_length to bigint in explore query
jsonb_array_length returns int4 but ExploreRouteRow expects i64 (int8),
causing sqlx type mismatch at runtime

+++ backend/routes/src/repository/postgres/mod.rs
@@ -195,7 +195,7 @@ impl RouteRepository for PostgresRouteRepository {
         let query = format!(
             r#"
             SELECT r.id, r.name,
-                   jsonb_array_length(r.points) AS points_count,
+                   jsonb_array_length(r.points)::bigint AS points_count,
                    r.created_at, r.share_token,
                    COALESCE(l.likes_count, 0) AS likes_count,
                    COALESCE(rt.avg_rating, 0.0) AS avg_rating,
